refactor(esl-utils): separate `esl-event-listeners` module
