+++ src/modules/esl-utils/dom/events.ts
@@ -1,4 +1,4 @@
 export * from './events/target';
 export * from './events/misc';
-export * from './events/listener';
-export * from './events/utils';
+
+export * from '../../esl-event-listener/core';
